spbp: The BP Based Survival Analysis Function

Semiparametric Survival Analysis Using Bernstein Polynomial

Usage

spbp(formula, ...)

Arguments

formula

a Surv object with time to event, status and explanatory terms.

...

Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.

Value

An object of class 'spbp'.

Details

Fits Bernstein Polynomial based Proportional regression to survival data.

See also

Examples

library("spsurv")
data("veteran", package = "survival")
#> Warning: data set ‘veteran’ not found

fit_mle <- spbp(Surv(time, status) ~ karno + factor(celltype),
  data = veteran, model = "po"
)
#> Priors are ignored because the MLE approach is used.
summary(fit_mle)
#> Bernstein Polynomial based Proportional Odds model
#> Call:
#> spbp.default(formula = Surv(time, status) ~ karno + factor(celltype), 
#>     data = veteran, model = "po", approach = "mle")
#> 
#>   n= 137, number of events= 128 
#> 
#>                               coef exp(coef) se(coef)     z Pr(>|z|)    
#> karno                     -0.06126   0.94058  0.00877 -6.98  2.9e-12 ***
#> factor(celltype)smallcell  1.28814   3.62603  0.43723  2.95   0.0032 ** 
#> factor(celltype)adeno      1.43747   4.21003  0.47367  3.03   0.0024 ** 
#> factor(celltype)large      0.10758   1.11358  0.46657  0.23   0.8176    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#>                           exp(coef) exp(-coef) lower .95 upper .95
#> karno                         0.941      1.063     0.925     0.957
#> factor(celltype)smallcell     3.626      0.276     1.539     8.543
#> factor(celltype)adeno         4.210      0.238     1.664    10.653
#> factor(celltype)large         1.114      0.898     0.446     2.779
#> 
#> Likelihood ratio test= 73.3  on 4 df,   p=5e-15
#> Wald test            = 65.4  on 4 df,   p=2e-13

fit_bayes <- spbp(Surv(time, status) ~ karno + factor(celltype),
  data = veteran, model = "po", approach = "bayes",
  cores = 1, iter = 300, chains = 1,
  priors = list(
    beta = c("normal(0,5)"),
    gamma = "halfnormal(0,5)"
  )
)
#> 
#> SAMPLING FOR MODEL 'spbp' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000439 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.39 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 1: Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 1: Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 1: Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 1: Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 1: Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 1: Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 1: Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 1: Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 1: Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 1: Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 1: Iteration: 300 / 300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2 seconds (Warm-up)
#> Chain 1:                1.3 seconds (Sampling)
#> Chain 1:                3.3 seconds (Total)
#> Chain 1: 
#> Warning: The largest R-hat is 1.08, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> 
#> Extracting posterior draws (this may take a moment)...

summary(fit_bayes)
#> Bayesian Bernstein Polynomial based Proportional Odds model
#> Call:
#> spbp.default(formula = Surv(time, status) ~ karno + factor(celltype), 
#>     data = veteran, approach = "bayes", model = "po", priors = list(beta = c("normal(0,5)"), 
#>         gamma = "halfnormal(0,5)"), cores = 1, chains = 1, iter = 300)
#> 
#>   n= 137, number of events= 128 
#> 
#>                           mean(coef) median(coef) mean(exp(coef)) sd(coef)
#> karno                       -0.05648     -0.05605         0.94512     0.01
#> factor(celltype)smallcell    1.07038      1.04843         3.16360     0.41
#> factor(celltype)adeno        1.19953      1.26197         3.64487     0.44
#> factor(celltype)large        0.00523      0.00499         1.10292     0.44
#> ---
#>                           mean(exp(coef)) mean(exp(-coef)) lower .95HPD
#> karno                               0.945            1.058        0.931
#> factor(celltype)smallcell           3.164            0.374        1.167
#> factor(celltype)adeno               3.645            0.331        1.434
#> factor(celltype)large               1.103            1.093        0.364
#>                           upper .95HPD
#> karno                             0.96
#> factor(celltype)smallcell         5.62
#> factor(celltype)adeno             7.21
#> factor(celltype)large             2.07
#> 
#> DIC=  1444   WAIC=  -720  LPML= -720